Add date-time picker documentation
